Move selenium-keepalive to lib folder
Cherry-picked from https://github.com/GooborgStudios/mdn-bcd-collector/commit/3287b5d9728472a38277b5cb509104ce8e2c5c0f.
